requirements.txt
torch==1.1.0
pyyaml
numpy